+++ src/features/setting/ui/passwordChangeModal.tsx
@@ -1,9 +1,7 @@
 import { useEffect } from "react";
 import { PasswordInput } from "@/entities/auth/ui";
-import { useSnackBar } from "@/shared/lib";
 import { useAuthStore } from "@/shared/store";
 import { Modal } from "@/shared/ui/modal";
-import { Snackbar } from "@/shared/ui/snackbar";
 import { usePutChangePassword } from "../api/putChangePassword";
 import { usePasswordChangeFormStore } from "../store";
 
@@ -114,11 +112,9 @@ export const PasswordChangeModal = ({
     (state) => state.reset,
   );
 
-  const { mutate: putChangePassword } = usePutChangePassword();
-  const { handleOpen: handleOpenSnackbar, onClose: onCloseSnackbar } =
-    useSnackBar(() => (
-      <Snackbar onClose={onCloseSnackbar}>비밀번호가 변경되었습니다.</Snackbar>
-    ));
+  const { mutate: putChangePassword } = usePutChangePassword({
+    onSuccessCallback: onClose,
+  });
 
   const handleSave = () => {
     const {
@@ -149,25 +145,12 @@ export const PasswordChangeModal = ({
       return;
     }
 
-    putChangePassword(
-      {
-        password: currentPassword,
-        newPw: newPassword,
-        newPwChk: confirmPassword,
-        token: token!,
-      },
-      {
-        onSuccess: () => {
-          resetPasswordChangeForm();
-          onClose();
-          handleOpenSnackbar();
-        },
-        onError: (error) => {
-          // TODO 에러 바운더리 로직 나오면 변경 하
-          console.error(error);
-        },
-      },
-    );
+    putChangePassword({
+      password: currentPassword,
+      newPw: newPassword,
+      newPwChk: confirmPassword,
+      token: token!,
+    });
   };
 
   useEffect(() => {
